Update download tests
Removes functionality for downloading from a static object: better to
remove this functionality than to reintroduce heavy-handed pruning.

If we want to re-add it, we can add a `get_pdf_url` accessor which
prefers a stored value on the struct to links... but it's nice to keep
the `Result` constructor simple.