Store Result.pdf_url member variable
Find the pdf_url during Result construction and store it. Deprecates the
get_pdf_url accessor.

It's unfortunate that we're duplicating data on the Result object again,
but it makes sense to move towards a more structured version of links
(e.g. indexing links in a dict by their content type) rather than
keeping a list.