Convert time.struct_time fields to datetimes
Issue: https://github.com/lukasschwab/arxiv.py/issues/57

This conversion was irritating in my work with the volt tutorial
tonight. This change is major enough to warrant a v1.1.0 release: it
changes the type (beyond correcting the annotation) of a field.

The package could keep the old field around and namespace access to the
new one, but I don't think v1.x.x adoption is widespread enough to merit
the technical debt.

TODO: add a unit test for the conversion.